refactor(console): Inbox
